add my::CubeTexture::DrawToSurface
git-svn-id: svn://127.0.0.1/repos/trunk@248 a5c22d9f-9f17-4845-9224-d5bed68bb274

+++ myd3dlib/src/myMesh.cpp
@@ -668,7 +668,7 @@ void OgreMesh::CreateMeshFromOgreXmlInMemory(
 	UnlockIndexBuffer();
 
 	std::vector<DWORD> rgdwAdjacency(GetNumFaces() * 3);
-	GenerateAdjacency(EPSILON_E6, &rgdwAdjacency[0]);
+	GenerateAdjacency((float)EPSILON_E6, &rgdwAdjacency[0]);
 	if(bComputeTangentFrame)
 	{
 		//DWORD dwOptions = D3DXTANGENT_GENERATE_IN_PLACE;
